aider: update `main/app.js` to resolve the error by requiring `connect-mongo` as a class and creating a new instance of `MongoStore` with the `client` property.
# Aider chat conversation:

USER: update `main/app.js` to resolve the following error: > [nodemon] starting `node app.js`
/Users/altninja/_sandbox/bengine/main/app.js:10
const MongoStore = require('connect-mongo')(session);
                                           ^

TypeError: Class constructor MongoStore cannot be invoked without 'new'
    at Object.<anonymous> (/Users/altninja/_sandbox/bengine/main/app.js:10:44)
    at Module._compile (node:internal/modules/cjs/loader:1254:14)
    at Module._extensions..js (node:internal/modules/cjs/loader:1308:10)
    at Module.load (node:internal/modules/cjs/loader:1117:32)
    at Module._load (node:internal/modules/cjs/loader:958:12)
    at Function.executeUserEntryPoint [as runMain] (node:internal/modules/run_main:81:12)
    at node:internal/main/run_main_module:23:47

ASSISTANT: To resolve the error, we need to modify the way `connect-mongo` is being required and used in `main/app.js`.

The error is due to the fact that `connect-mongo` exports a class, and we are trying to invoke it as a function. We need to create a new instance of `MongoStore` using the `new` keyword.

Here is the *SEARCH/REPLACE* block:

main/app.js
```javascript
<<<<<<< SEARCH
const MongoStore = require('connect-mongo')(session);
=======
const MongoStore = require('connect-mongo').default;
>>>>>>> REPLACE
```

main/app.js
```javascript
<<<<<<< SEARCH
store: new MongoStore({ mongooseConnection: mongoose.connection }),
=======
store: new MongoStore({ client: mongoose.connection.getClient() }),
>>>>>>> REPLACE
```

In the first block, we change the way we require `connect-mongo` to get the default export, which is the `MongoStore` class.

In the second block, we change the way we create a new `MongoStore` instance. We pass an object to the constructor with a `client` property, which should be a MongoDB client instance. We get this from the mongoose connection using the `getClient()` method.